requirements.txt
torch
fair-esm
streamlit
numpy
tqdm
